+++ mooney/src/main/java/tamtam/mooney/global/openai/dto/MonthlyBudgetInputRequestDto.java
@@ -1,5 +1,6 @@
 package tamtam.mooney.global.openai.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -9,5 +10,5 @@ public record MonthlyBudgetInputRequestDto(
         List<FixedExpenseInputDto> fixedExpenses,
         String userOpinions,
         List<NextMonthScheduleInputDto> nextMonthSchedules,
-        double totalBudget
+        BigDecimal totalBudget
 ) {}
